[patch] Correct catalog manifests
A bug generate-manifests has resulted in string based comparison of version strings, rather than a semver aware comparison.  This results in versions such as "1.0.9" appearing to be greater than "1.0.19" for instance, and the wrong version being listed in the manifest.

The manifest for all catalogs released since June 2023 have been updated with the correct manifest.